AppVeyor: Fix creation of symblic links for good
The main repository contains symbolic links to directories in the
submodule. As the submodule is initialized after the main repository, it
is unknown to Git whether the links point to files or directories. On
Windows, this is important to know as symbolic links can point to
either files or directories, and the type has to be known at the point of
their creation. The default is to create file symbolic links, which is
wrong in this case. So delete the symbolic links and check them out
again after the submodule has been initialized to get them recreated as
directory symbolic links.