Remove Result wrappers on some methods
`analyze` never had a case where it could fail. Introducing a failure
would be a breaking change in behavior, so I'd rather alter the type
if that ever needs to happen.

The TextData methods could all return Err in the event of a programming
error, but these have been changed to panic instead. There wasn't a
reasonable case where they'd return Err due to user input.