Fix remove_common_tokens
This is particularly diffucult, as the example that this LCS removal
enables is one of Adobe's "%%Copyright" prefixes. Guess what else has
Copyright at the start of a line? Yep, actual copyright statements.

Fixed some inconsistencies within remove_common_tokens and
lcs_substring, and additionally changed overall logic:

Track the counts of common prefixes and store them. Take the most common
and re-count against other found prefixes.

If that count is at least 80% of the text, consider it to be a junk line
prefix and remove it.

There are probably better ways to do this, but this fits the bill for
places where licenses tend to appear.

Fixes #42