Add --multiple option to the CLI
The --multiple option enables uses ScanMode::TopDown to detect multiple
licenses in the same file. The confidence score is still zero for the
whole file but having a list of licenses in the file is still useful. An
improvement would be to use the individual confidence scores in the
reported total score.

Added a test case that uses the license from python-zeep that issue #40
is about where the file has one MIT and two BSD-3-Clause licenses in it.